requirements.txt
Django == 3.1.1
gunicorn == 20.0.4
djangorestframework == 3.11.1
django-oauth-toolkit == 1.3.2
django-rest-framework-social-oauth2 == 1.1.0
django-cors-headers == 3.5.0
firebase-admin ==  4.4.0
pyrebase4 == 4.4.1